- backend: model_file : replace get_file_timestamp with base::file_mtime - compilation problem in OEL - grtpp, dbquery: temporary add DEFINE_INT_FUNCTIONS to fix compile issues in OEL 32bit - will be fixed after merging biginteger patch - dbquery: switch uint64_t to long - it was causing compile problems in OEL32bit - will be fixed after merging biginteger patch [TESTS]  - checked if wb properly builds on ubuntu 12.04 64bit, oel 64bit, oel 32bit

+++ frontend/linux/workbench/overview_panel.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2007, 2013, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2007, 2014, Oracle and/or its affiliates. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -825,7 +825,7 @@ public:
 private:
   void page_switched(GtkNotebookPage *page, guint num)
   {
-    if ((int)num < _overview->count_children(_node))
+    if ((size_t)num < _overview->count_children(_node))
       _overview->focus_node(_overview->get_child(_node, num));
     _current_page_index= num;
   }
